Move message manipulation methods into a new class
As described by https://github.com/UST-MICO/mico/issues/762 we want
to split up our logic. This moves the message manipulation related
methods from MessageListener and KafkaMessageSender to CloudEventManipulator.
